demos: fix descriptor passing, add selecty
Fix a discrepency between control message _space_ and _length_
that caused too many descriptors to be passed between processes.

Also get rid of heap allocation by using the C11 `alignas` type
qualifier.

Add a demo that shows how to use `select` on descriptors passed
from a dedicated listener; this involves non-blocking IO: since
all readers of the listener socket become ready when the
listener writes to its end, they all call `recvmsg` when a new
connection is passed across the socket.  But only one can win
the race to actually receive the new descriptor, so the rest
end up blocking in `recvfd`, starving out the descriptors that
they have been passed.  To work around that, we set the read end
of the listener socket to be non-blocking.  In this case, all
worker processes will still see the listener socket become ready
for reading when the listener sends a descriptor, but still only
one will win the race and actually receive the new socket; the
rest will return an error with `errno` set to `EWOULDBLOCK`,
which we can detect and ignore.

Signed-off-by: Dan Cross <cross@gajendra.net>